update: remove task module and refactore shared folders schema
