+++ src/main/java/emerick/igor/javatodolist/shared/infra/config/FiltersConfig.java
@@ -1,4 +1,4 @@
-package emerick.igor.javatodolist.shared.config;
+package emerick.igor.javatodolist.shared.infra.config;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 import emerick.igor.javatodolist.modules.user.database.repositories.models.IUserRepository;
-import emerick.igor.javatodolist.shared.filters.AuthenticationFilter;
+import emerick.igor.javatodolist.shared.infra.filters.AuthenticationFilter;
 
 @Configuration
 public class FiltersConfig {
