+++ src/main/java/emerick/igor/javatodolist/shared/infra/errors/handlers/ErrorHandler.java
@@ -1,10 +1,10 @@
-package emerick.igor.javatodolist.shared.errors.handlers;
+package emerick.igor.javatodolist.shared.infra.errors.handlers;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 
-import emerick.igor.javatodolist.shared.errors.HttpError;
+import emerick.igor.javatodolist.shared.infra.errors.HttpError;
 
 @ControllerAdvice
 public class ErrorHandler {
